reveal-presentation-core/tools/requirements_ppt.txt
pdf2image>=1.16.0
python-pptx>=0.6.21
Pillow>=9.0.0
